// Frontend/katana-slice-manager/src/app/models/interfaces/proxmox-vm-creation-form.interface.ts
export interface ProxmoxVmTargetFormModel {
  node: string;
  vmName: string;
  template: string;
  isoImage: string;
  start: boolean;
  cpu: number;
  ram: number;
  storageType: string;
  diskSize: number;
}

export interface ProxmoxVmCreationFormModel {
  clusterName: string;
  vmName: string;
  template: string;
  isoImage: string;
  start: boolean;
  cpu: number;
  ram: number;
  storageType: string;
  diskSize: number;
  vmTargets: ProxmoxVmTargetFormModel[];
  managementBridgeName: string;
  managementBridgeType: string;
  customBridgeName: string;
  customBridgeType: string;
  customIp: string;
  customNetmask: string;
  customGateway: string;
}